+++ 3_catmark_subdiv/mainwindow.h
@@ -26,7 +26,7 @@ class MainWindow : public QMainWindow {
   void on_MeshPresetComboBox_currentTextChanged(const QString &meshName);
   void on_SubdivSteps_valueChanged(int subdivLevel);
   void on_TessellationCheckBox_toggled(bool checked);
-  void on_EdgeSharpness_valueChanged(int sharpness);
+  void on_EdgeSharpness_valueChanged(double sharpness);
 
   void on_HideMeshCheckBox_toggled(bool checked);
   void on_LimitPositionCheckBox_toggled(bool checked);
@@ -35,7 +35,7 @@ class MainWindow : public QMainWindow {
   void on_BezierRadio_toggled(bool checked);
   void on_BSplineRadio_toggled(bool checked);
   
-  void onEdgeSelected(int sharpness);  // Slot for edge selection signal
+  void onEdgeSelected(float sharpness);  // Slot for edge selection signal
   void onVertexSelected(int sharpEdgeCount);  // Slot for vertex selection signal
 
  private:
